Revert "refactor/Devtooling-186: Routing Queue  (#1213)"
This reverts commit a520fd9004ae3923e07b006b48796333a8de87bd.
