+++ examples/resources/genesyscloud_flow/inboundcall_flow_example.yaml
@@ -1,5 +1,5 @@
 inboundCall:
-  name: Terraform Flow Test-b066d719-c330-41c3-8c0f-9f880f6cdfb5
+  name: Terraform Flow Test-a8bf377f-3900-42a2-aa43-cec584f14366
   defaultLanguage: en-us
   startUpRef: ./menus/menu[mainMenu]
   initialGreeting:
