+++ examples/resources/genesyscloud_flow/inboundcall_flow_example2.yaml
@@ -1,5 +1,5 @@
 inboundEmail:
-    name: Terraform Flow Test-88aecd38-8ab8-45cf-bf74-7f9fa01da2ac
+    name: Terraform Flow Test-c93ae359-ace8-4e15-932b-18f242a33b09
     division: New Home
     startUpRef: "/inboundEmail/states/state[Initial State_10]"
     defaultLanguage: en-us
