+++ examples/resources/genesyscloud_flow/inboundcall_flow_example3.yaml
@@ -1,5 +1,5 @@
 inboundCall:
-  name: Terraform Flow Test-bbe3c4a2-faa9-48c7-a54d-6117ae2eabc1
+  name: Terraform Flow Test-7fbe9a86-3aca-4bdc-a88f-ab89b891793e
   defaultLanguage: en-us
   startUpRef: ./menus/menu[mainMenu]
   initialGreeting:
