+++ genesyscloud/routing_queue/data_source_genesyscloud_routing_queue.go
@@ -15,6 +15,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/mypurecloud/platform-client-sdk-go/v133/platformclientv2"
 )
 
 var (
@@ -32,6 +33,7 @@ func dataSourceRoutingQueueRead(ctx context.Context, d *schema.ResourceData, m i
 	}
 
 	queueId, err := rc.RetrieveId(dataSourceRoutingQueueCache, resourceName, key, ctx)
+
 	if err != nil {
 		return err
 	}
@@ -48,22 +50,34 @@ func normalizeQueueName(queueName string) string {
 // hydrateRoutingQueueCacheFn for hydrating the cache with Genesys Cloud routing queues using the SDK
 func hydrateRoutingQueueCacheFn(c *rc.DataSourceCache) error {
 	log.Printf("hydrating cache for data source genesyscloud_routing_queues")
-	proxy := GetRoutingQueueProxy(c.ClientConfig)
-	ctx := context.Background()
+	routingApi := platformclientv2.NewRoutingApiWithConfig(c.ClientConfig)
+	const pageSize = 100
+	queues, _, getErr := routingApi.GetRoutingQueues(1, pageSize, "", "", nil, nil, nil, "", false)
 
-	allQueues, resp, err := proxy.GetAllRoutingQueues(ctx, "")
-	if err != nil {
-		return fmt.Errorf("failed to get routing queues %s %s", err, resp)
+	if getErr != nil {
+		return fmt.Errorf("failed to get page of skills: %v", getErr)
 	}
-
-	if allQueues == nil || len(*allQueues) == 0 {
+	if queues.Entities == nil || len(*queues.Entities) == 0 {
 		return nil
 	}
-
-	for _, queue := range *allQueues {
+	for _, queue := range *queues.Entities {
 		c.Cache[normalizeQueueName(*queue.Name)] = *queue.Id
 	}
 
+	for pageNum := 2; pageNum <= *queues.PageCount; pageNum++ {
+
+		queues, _, getErr := routingApi.GetRoutingQueues(pageNum, pageSize, "", "", nil, nil, nil, "", false)
+		if getErr != nil {
+			return fmt.Errorf("failed to get page of queues: %v", getErr)
+		}
+		if queues.Entities == nil || len(*queues.Entities) == 0 {
+			break
+		}
+		// Add ids to cache
+		for _, queue := range *queues.Entities {
+			c.Cache[normalizeQueueName(*queue.Name)] = *queue.Id
+		}
+	}
 	log.Printf("cache hydration completed for data source genesyscloud_routing_queues")
 	return nil
 }
@@ -71,20 +85,27 @@ func hydrateRoutingQueueCacheFn(c *rc.DataSourceCache) error {
 // Get queue by name.
 // Returns the queue id (blank if not found) and diag
 func getQueueByNameFn(c *rc.DataSourceCache, name string, ctx context.Context) (string, diag.Diagnostics) {
-	proxy := GetRoutingQueueProxy(c.ClientConfig)
+	routingApi := platformclientv2.NewRoutingApiWithConfig(c.ClientConfig)
 	queueId := ""
-
+	const pageSize = 100
 	diag := util.WithRetries(ctx, 15*time.Second, func() *retry.RetryError {
-		queueID, resp, retryable, getErr := proxy.getRoutingQueueByName(ctx, name)
-		if getErr != nil && !retryable {
-			return retry.NonRetryableError(util.BuildWithRetriesApiDiagnosticError(resourceName, fmt.Sprintf("error requesting queue %s | error %s", name, getErr), resp))
-		}
-		if retryable {
-			return retry.NonRetryableError(util.BuildWithRetriesApiDiagnosticError(resourceName, fmt.Sprintf("error requesting queue %s | error %s", name, getErr), resp))
-		}
+		for pageNum := 1; ; pageNum++ {
+			queues, resp, getErr := routingApi.GetRoutingQueues(pageNum, pageSize, "", name, nil, nil, nil, "", false)
+			if getErr != nil {
+				return retry.NonRetryableError(util.BuildWithRetriesApiDiagnosticError(resourceName, fmt.Sprintf("error requesting queue %s | error %s", name, getErr), resp))
+			}
 
-		queueId = queueID
-		return nil
+			if queues.Entities == nil || len(*queues.Entities) == 0 {
+				return retry.RetryableError(util.BuildWithRetriesApiDiagnosticError(resourceName, fmt.Sprintf("no routing queues found with name %s", name), resp))
+			}
+
+			for _, queue := range *queues.Entities {
+				if queue.Name != nil && normalizeQueueName(*queue.Name) == normalizeQueueName(name) {
+					queueId = *queue.Id
+					return nil
+				}
+			}
+		}
 	})
 
 	return queueId, diag
